Refactor common module to avoid setuptools import
We do not need the setuptools import used while building packages in
ppbt.common. Having that import around breaks relenv runtime when trying
to setup the toolchain.
